Dangerous.jl

Dangerous.Hamiltonian.σ_x — Constant

σ_x

Single-spin Pauli matrix.

\[\sigma_x = \frac{1}{2}\begin{pmatrix}0 & 1 \\ 1 & 0\end{pmatrix}\]

Dangerous.Hamiltonian.σ_y — Constant

σ_y

Single-spin Pauli matrix.

\[\sigma_y = \frac{1}{2}\begin{pmatrix}0 & -\mathrm{i} \\ \mathrm{i} & 0\end{pmatrix}\]

Dangerous.Hamiltonian.σ_z — Constant

σ_z

Single-spin Pauli matrix.

\[\sigma_z = \frac{1}{2}\begin{pmatrix}1 & 0 \\ 0 & -1\end{pmatrix}\]

Dangerous.Hamiltonian.h_coupling — Method

h_coupling(sys)

Generate the spin–spin coupling Hamiltonian for a spin system, given by:

\[H_{\text{coupling}} = \sum_{i \neq j} 2\pi J_{ij} (I_{ix}I_{jx} + I_{iy}I_{jy} + I_{iz}I_{jz})\]

where $J_{ij}$ is the coupling constant between spins i and j (in Hz).

Dangerous.Hamiltonian.h_free — Method

h_free(sys)

Generate the free Hamiltonian for a spin system.

\[H_{\text{free}} = H_{\text{offset}} + H_{\text{coupling}}\]

Dangerous.Hamiltonian.h_offset — Method

h_offset(sys)

Generate the offset Hamiltonian for a spin system, given by:

\[H_{\text{offset}} = \sum_i \Omega_{0,i} I_{iz}\]

where $\Omega_{0,i} = -\gamma_i B_0 - \omega_{\text{ref},i}$ is the offset angular frequency of spin i.

Dangerous.Hamiltonian.h_pulse — Method

h_pulse(sys, nuc, ω1, φ)

Generate a Hamiltonian for a pulse on nucleus nuc, with angular frequency (ω1 = B1/γ) and pulse phase φ.

The angular frequency should be specified in rad/s. The pulse phase can be either specified as an angle in radians, or as one of the following symbols: :x, :y, :_x, or :_y, which correspond respectively to 0, π/2, π, and 3π/2.

Dangerous.Hamiltonian.op1 — Method

op1(i, n, single_spin_op)

Generate a single-spin operator acting on spin i in a system of n spins.

e.g. $I_{1z}$ in a 4-spin system is generated by op1(1, 4, σ_z).

Dangerous.Hamiltonian.op2 — Method

op2(i, j, n, single_spin_op1, single_spin_op2)

Generate a two-spin operator between spins i and j in a system of n spins.

e.g. $I_{1z}I_{2x}$ in a 4-spin system is generated by op2(1, 2, 4, σ_z, σ_x).

Dangerous.Hamiltonian.ρ_eq — Method

ρ_eq(sys)

Generate the thermal density matrix for a spin system. Formally, this is given by

\[\rho_{\text{eq}} = \frac{\exp{(-\beta H)}}{\text{Tr}[(\exp{(-\beta H)]}}\]

where $\beta = 1/k_\mathrm{B}T$, but here we use the approximation of

\[\rho_{\text{eq}} \propto \sum_i \gamma_i I_{iz}.\]

This is justified by the fact that $H$ is very small (so $\exp(-\beta H) \approx 1 - \beta H$) and that the coupling term is small compared to the offset term.

Dangerous.SpinSystem.System — Type

System(
    magnetic_field::MagneticField,
    nuclei::Vector{Nucleus},
    chemical_shifts::Vector{<:Real},
    couplings::Matrix{Unitful.Frequency}
)

A complete specification of a spin system.

TODO: The magnetic field is not part of the physical spin system, but rather the spectrometer / experiment setup. It should logically be moved elsewhere.

Fields:

magnetic_field::MagneticField: The magnetic field strength, e.g. 22.1u"T".

nuclei::Vector{Nucleus}: The nuclei in the system.

chemical_shifts::Vector{Float64}: The chemical shifts of the nuclei in the system, in ppm. The length of this array must match the length of nuclei.

couplings::Matrix{Float64}: The coupling constants between the nuclei in the system. The size of this matrix must be (length(nuclei), length(nuclei)). Note that the diagonal elements of this matrix must be zero, and for each pair of off-diagonal elements (i, j) and (j, i), one of them must be zero. The easiest way to generate this matrix is to initialise it with zeros(n, n), then set the non-zero elements.