Dangerous.jl

Dangerous.Hamiltonian.op1 — Method

op1(i, n, single_spin_op)

Generate a single-spin operator acting on spin i in a system of n spins.

e.g. $I_{1z}$ in a 4-spin system is generated by $op1(1, 4, SingleSpin.Z)$.

Dangerous.Hamiltonian.op2 — Method

op2(i, j, n, single_spin_op1, single_spin_op2)

Generate a two-spin operator between spins i and j in a system of n spins.

e.g. $I_{1z}I_{2x}$ in a 4-spin system is generated by $op2(1, 2, 4, SingleSpin.Z, SingleSpin.X)$.

Dangerous.Hamiltonian.SingleSpin.X — Constant

X, Y, Z

Single-spin Pauli matrices.

Dangerous.Hamiltonian.SingleSpin.h_offset — Method

h_offset(Ω)

Generate a Hamiltonian for a spin offset.

The offset frequency should be specified in Hz.

Dangerous.Hamiltonian.SingleSpin.h_pulse — Method

h_pulse(ω1, φ)

Generate a Hamiltonian for a pulse with a given angular frequency (ω1 = B1/γ) and pulse phase.

The angular frequency should be specified in rad/s. The pulse phase can be either specified as an angle in radians, or as one of the following symbols: :x, :y, :_x, or :_y, which correspond respectively to 0, π/2, π, and 3π/2.